requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
APScheduler==3.11.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
build==1.3.0
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.2
chromedriver==2.24.1
colorama==0.4.6
configparser==7.2.0
coverage==7.10.2
crayons==0.4.0
Django==5.2.5
django-apscheduler==0.7.0
django-extensions==4.1
django-rest-framework==0.1.0
django-timezone-field==7.1
djangorestframework==3.16.1
entrypoints==0.4
environ==1.0
flake8==7.3.0
frozenlist==1.7.0
furl==2.1.4
future==1.0.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
meraki==2.0.3
more-itertools==10.7.0
multidict==6.6.3
nose==1.3.7
orderedmultidict==1.0.1
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
py==1.11.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pyngrok==7.3.0
pyparsing==3.2.3
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==8.4.1
pytest-django==4.11.1
pytest-html==4.1.1
pytest-metadata==3.1.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
toml==0.10.2
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
